(fix) Synchronize sas token test completion.
