+++ e2etests/test/device_acknowledge_tests.js
@@ -7,6 +7,7 @@ var serviceSdk = require('azure-iothub');
 var createDeviceClient = require('./testUtils.js').createDeviceClient;
 var closeDeviceServiceClients = require('./testUtils.js').closeDeviceServiceClients;
 var DeviceIdentityHelper = require('./device_identity_helper.js');
+var Rendezvous = require('./rendezvous_helper.js').Rendezvous;
 
 var assert = require('chai').assert;
 var debug = require('debug')('e2etests');
@@ -17,35 +18,6 @@ var deviceAmqp = require('azure-iot-device-amqp');
 
 var hubConnectionString = process.env.IOTHUB_CONNECTION_STRING;
 
-function Rendezvous(done) {
-  this.doneYet = {};
-  this.done = done;
-  this.everybodyDone = true;
-}
-
-Rendezvous.prototype.imIn = function(participant) {
-  if (this.doneYet.hasOwnProperty(participant)) {
-    throw new Error('can not participate more than once');
-  }
-  this.doneYet[participant] = false;
-};
-
-Rendezvous.prototype.imDone = function(participant) {
-  if (Object.keys(this.doneYet).length === 0) {
-    throw new Error('Nobody joined to rendezvous');
-  }
-  if (this.doneYet[participant]) {
-    throw new Error('participant can not say done more than once');
-  }
-  this.doneYet[participant] = true;
-  this.everybodyDone = true;
-  Object.keys(this.doneYet).forEach(function(aParticipant) {
-    this.everybodyDone = this.everybodyDone && this.doneYet[aParticipant];
-  }.bind(this));
-  if (this.everybodyDone) {
-    return this.done();
-  }
-};
 [
   DeviceIdentityHelper.createDeviceWithSas,
   DeviceIdentityHelper.createDeviceWithSymmetricKey,
